+++ test/fixtures/with-deps/task-1.js
@@ -1,0 +1,5 @@
+export default function factory() {
+    return function task1(done) {
+        done();
+    };
+}