+++ test/fixtures/with-deps/task-2.js
@@ -1,0 +1,5 @@
+export default function factory() {
+    return function task2(done) {
+        done();
+    };
+}